+++ packages/stream_chat_flutter/lib/src/attachment/url_attachment.dart
@@ -79,7 +79,7 @@ class StreamUrlAttachment extends StatelessWidget {
                             borderRadius: const BorderRadius.only(
                               topRight: Radius.circular(16),
                             ),
-                            color: messageTheme.linkBackgroundColor,
+                            color: messageTheme.urlAttachmentBackgroundColor,
                           ),
                           child: Padding(
                             padding: const EdgeInsets.only(
@@ -89,9 +89,7 @@ class StreamUrlAttachment extends StatelessWidget {
                             ),
                             child: Text(
                               hostDisplayName,
-                              style: chatThemeData.textTheme.bodyBold.copyWith(
-                                color: chatThemeData.colorTheme.accentPrimary,
-                              ),
+                              style: messageTheme.urlAttachmentHostStyle,
                             ),
                           ),
                         ),
@@ -109,14 +107,12 @@ class StreamUrlAttachment extends StatelessWidget {
                         urlAttachment.title!.trim(),
                         maxLines: messageTheme.urlAttachmentTitleMaxLine ?? 1,
                         overflow: TextOverflow.ellipsis,
-                        style: chatThemeData.textTheme.body
-                            .copyWith(fontWeight: FontWeight.w700),
+                        style: messageTheme.urlAttachmentTitleStyle,
                       ),
                     if (urlAttachment.text != null)
                       Text(
                         urlAttachment.text!,
-                        style: chatThemeData.textTheme.body
-                            .copyWith(fontWeight: FontWeight.w400),
+                        style: messageTheme.urlAttachmentTextStyle,
                       ),
                   ],
                 ),